Change min-tmpfs to 3
Setting it to 4 would prevent it from using tmpfs even on my 8 GB
VM because the tmpfs size there is set to 3.9 GB (which works fine).
Setting the minimum to 3 will still prevent it from being used on
my 6 GB system because that has a tmpfs size of 2.8 GB (which does
not work).